+++ addons/addons.go
@@ -25,5 +25,4 @@ import (
 	_ "github.com/invopop/gobl/addons/it/ticket"
 	_ "github.com/invopop/gobl/addons/mx/cfdi"
 	_ "github.com/invopop/gobl/addons/pl/favat"
-	_ "github.com/invopop/gobl/addons/pt/saft"
 )