+++ addons/external.go
@@ -66,6 +66,13 @@ func init() {
 		},
 		Module: "github.com/invopop/gobl.sa.zatca",
 	})
+  tax.RegisterApprovedAddon(&tax.ExternalAddon{
+		Key: "pt-saft-v1",
+		Name: i18n.String{
+			i18n.EN: "Portugal SAF-T",
+		},
+		Module: "github.com/invopop/gobl.pt.saft",
+	})
 	tax.RegisterApprovedAddon(&tax.ExternalAddon{
 		Key: "br-nfe-v4",
 		Name: i18n.String{
@@ -79,5 +86,5 @@ func init() {
 			i18n.EN: "Brazil NFS-e 1.X",
 		},
 		Module: "github.com/invopop/gobl.br.nfse",
-	})
+  })
 }